ログイン認証処理追加

+++ src/SequenceLogService/src/getSequenceLogListJson.cpp
@@ -107,7 +107,7 @@ static void createSequenceLogListJson(Json* json, FileInfo* info)
 
         if (size < 1024)
         {
-            strSize.format("%s %d byte(s)", message.getBuffer(), (int64_t)size);
+            strSize.format("%s %d byte(s)", message.getBuffer(), (int32_t)size);
         }
 
         else
